programmed the current features
coded the features in the flowchar to ensure program viability. Added a few required features missing in flowchart such as specific colours for the texts and canvas, the mainloop() statement. Going back to the flowchart to add these amendments to the program. Will need to make one more test on the functionality of radiobuttons and checker buttons to ensure find out how to apply their function into the actual program. Flowchart will only be updated with the specifics on how to gather data from the radio and checks, once a beta configuration has been tested in a future verson of the test.py file.