Basic repeat process of planning then coding for a few days.
Toyed around with the idea of using a custom class to create buttons. conclusion is that it is simply not worth the complication nor the additional lines of code. Yes, it does make establishing buttons easier, and will make future changes to the program easier. However, as each of the button's functions are so varied, and those functions have to be individually coded, and the inputs seperatly handled, it does not simplfy things to a great deal. It was decided that the additional code was not worth the code saving, or the ease of use gained.

Continuing on with planning 2. Will not need test file anymore